src/backend/requirements.txt

# Django and REST Framework requirements
django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
requests==2.31.0
pandas==2.1.4
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
